requirements.txt
pyaudio
memory-profiler
scipy
requests_async
# INSTALL TENSORFLOW https://www.youtube.com/watch?v=vekblEk6UPc
